requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
idna-ssl==1.1.0
multidict==4.7.4
pycparser==2.19
python-telegram-bot==12.3.0
requests==2.22.0
selenium==3.141.0
six==1.14.0
telepot==12.7
tornado==6.0.3
typing-extensions==3.7.4.1
urllib3==1.25.7
yarl==1.4.2
